+++ CHANGELOG.md
@@ -10,8 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Categories list page
 - Categories details page
 - Set category color
-- Add priority switch in habit, for more prominent display
-- Add priority icon in habits list card
+- Priority switch in habit, for more prominent display
+- Priority icon in habit settings list card
+- Priority icon in habits list card
 
 ### Changed:
 - Whitespace in settings